Give more insight into the "magic" of useReducer
As a deep dive blog post, it's helpful to demystify any of the magic behind counter-intuitive results. The "cheat mode" section I think merited a bit better explanation of how useReducer magically is aware of the latest props. I'm not 100% sure if my pseudo implementation actually is accurate but it's the best mental model I was able to come up with after reading this post and glancing through the React source. If it's inaccurate or works through a completely different implementation I'd love to see it :)