useEffect guide: modify `someone`'s name property
The blog post contains a pure-JavaScript example of caching a local `name` value for later reuse in `setTimeout`. However, when the whole `someone` variable is reassigned before each call, the version with `setTimeout` taking `name` directly from `person`:
```js
function sayHi(person) {
  setTimeout(() => {
    alert('Hello, ' + person.name);
  }, 3000);
}
```
would work the same way. To illustrate the difference, it'd be better to modify the `name` property instead which I did in this PR.

I also applied the modifications to the CodeSandbox and swapped the link. I assume you might want to fork that one back or re-apply the modifications by yourself.