Update link to ReactFiberHooks source
Hi, Dan. I was reading you great article “React as a UI runtime” and noticed the link to the source of ReactFiberHooks is broken right now as it’s missing the old/new suffix. 

Btw, I assumed it was the old version because the new one was created after your article. Is this correct?